add model and func

+++ app/model/file_folder.go
@@ -3,7 +3,7 @@ package model
 type FileFolder struct {
 	BaseModel
 	FileFolderName string
-	ParentFolderId int64
-	FileStoreId    int64
+	ParentFolderId int
+	FileStoreId    int
 	Time           string
 }
